backend/auth/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
bcrypt==3.2.2
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-mail==1.1.4
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
motor==3.0.0
orjson==3.7.7
packaging==21.3
passlib==1.7.4
prompt-toolkit==3.0.30
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
cloudinary==1.30.0